+++ src/jsv_type_ksuid.erl
@@ -16,7 +16,7 @@
 
 -behaviour(jsv_type).
 
--export([validate_type/1, canonicalize/3]).
+-export([validate_type/1, canonicalize/3, generate/2]).
 
 validate_type(Value) when is_binary(Value) ->
   case ksuid:parse(Value) of
@@ -30,3 +30,8 @@ validate_type(_) ->
 
 canonicalize(_, CData, _) ->
   CData.
+
+generate(Term, _) when is_binary(Term), byte_size(Term) =:= 20 ->
+  {ok, ksuid:format(Term)};
+generate(_, _) ->
+  invalid.