+++ src/jsv_type_uri.erl
@@ -16,7 +16,7 @@
 
 -behaviour(jsv_type).
 
--export([validate_type/1]).
+-export([validate_type/1, generate/2]).
 
 validate_type(<<"">>) ->
   error;
@@ -29,3 +29,8 @@ validate_type(Value) when is_binary(Value) ->
   end;
 validate_type(_) ->
   error.
+
+generate(Term, _) when is_binary(Term) ->
+  {ok, Term};
+generate(_, _) ->
+  invalid.