+++ test/jsv_validator_tests.erl
@@ -12,7 +12,7 @@
 %% OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
 %% PERFORMANCE OF THIS SOFTWARE.
 
--module(jsv_test).
+-module(jsv_validator_tests).
 
 -include_lib("eunit/include/eunit.hrl").